moving common function out of bound vectors
